requirements.txt
sqlalchemy
psycopg2-binary
#psycopg2
pgvector

alembic

fastapi
uvicorn
pydantic[email]
pydantic-settings
python-multipart

bcrypt
pyjwt

pymupdf
python-docx
langchain
langchain-community
langchain-text-splitters
#datetime
sentence-transformers
torch

openai
anthropic
google-generativeai
groq

cryptography

python-dotenv

streamlit
requests

